requirements.txt
asgiref==3.5.2
attrs==21.4.0
black==22.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
coreapi==2.3.3
coreapi-cli==1.0.9
coreschema==0.0.4
coverage==6.4.2
Django==4.0.6
django-environ==0.9.0
djangorestframework==3.13.1
idna==3.3
importlib-metadata==4.12.0
iniconfig==1.1.1
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.4.1
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.5.2
pluggy==1.0.0
py==1.11.0
Pygments==2.12.0
PyJWT==2.4.0
pyparsing==3.0.9
pytest==7.1.2
pytest-cov==3.0.0
pytest-django==4.5.2
pytest-env==0.6.2
pytz==2022.1
requests==2.28.1
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.3.0
uritemplate==4.1.1
urllib3==1.26.10
zipp==3.8.1
